% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spec_tools.R
\name{bin_expression_data}
\alias{bin_expression_data}
\title{Compute bins for expression data}
\usage{
bin_expression_data(expression_data, method = "Efroni", ...)
}
\arguments{
\item{expression_data}{data frame (or tibble) containing the following
columns for a single locus

\itemize{
  \item Expression (the normalized expression value, not log-transformed)
  \item Cell_Type (the cell type from which the expression data came from)
}}

\item{method}{a character (possible values consist only of "Efroni" for now)
OR a function that takes in the `expression_data` data.frame object and any
other parameters, and returns expression_data with an extra column, "bin"
containing the expression bin for that dataset/cell_type combination}

\item{...}{Additional parameters to the `method` function argument, if specified}
}
\value{
a data.frame (or tibble) containing the data within `expression_data`
and an additional "bin" column describing the expression bin for that
expression/cell_type combination.
}
\description{
takes in expression and cell-type data for a single locus, and
computes expression bins.
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification of cell identity from single-cell gene expression profiles". Genome Biology 16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy biological systems". Nucl. Acids. Res. 39(21)
}
