% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{compute_ici_for_cell}
\alias{compute_ici_for_cell}
\title{Compute ici score for all cell types for a given cell's expression profile}
\usage{
compute_ici_for_cell(expression_and_marker_data, sig = FALSE,
  n_iterations = 1000)
}
\arguments{
\item{expression_and_marker_data}{data.frame with the following columns for
each cell type (melted):
\itemize{
  \item Locus, the gene name
  \item Cell_Type, the cell type for each Locus/Expression/spec/is_marker combination
  \item is_marker, logical, whether each gene is a marker
  \item Expression, numeric, the expression of each locus for this profile
  \item spec, numeric, the specificity score for this each locus for the cell
  type in question.
 }}

\item{sig}{logical, whether to compute p-values for ici scores}

\item{n_iterations}{integer, how many random permutations for p-value calculations}
}
\value{
a data.frame with three columns:
\itemize{
  \item Cell_Type, the cell type for each Locus/Expression/spec/is_marker combination
  type in question.
  \item ici_score the ici score for that cell/cell-type
  \item p-val the p-value associated with the ici-score, NA if sig = FALSE.
 }
}
\description{
Compute ici score for all cell types for a given cell's expression profile
}
