% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{compute_ici_pval}
\alias{compute_ici_pval}
\title{Compute p-values for a single expression/cell type profile (Internal)}
\usage{
compute_ici_pval(cell_type_data, ici_score = NULL, n_iterations = 1000)
}
\arguments{
\item{cell_type_data}{a data.frame with the following columns:
\itemize{
  \item Locus, the gene name
  \item is_marker, logical, whether each gene is a marker
  \item Expression, numeric, the expression of each locus for this profile
  \item spec, numeric, the specificity score for this each locus for the cell
  type in question.
 }}

\item{ici_score}{numeric, the ici score to compare the random iterations against.
If null (default), will compute denovo}

\item{n_iterations, }{integer, number of random iterations}
}
\value{
numeric, raw p-value indicating the number of randomized ici scores
that are greater than or equal to the "real" ici score.
}
\description{
Compute p-values for a single expression/cell type profile (Internal)
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification
of cell identity from single-cell gene expression profiles". Genome Biology
16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy
biological systems". Nucl. Acids. Res. 39(21)
}
