% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{compute_ici_score_for_cell_type}
\alias{compute_ici_score_for_cell_type}
\title{Compute ICI score for a given cell type (Internal)}
\usage{
compute_ici_score_for_cell_type(marker_expression, marker_spec)
}
\arguments{
\item{marker_expression}{numeric vector, gene expression profile}

\item{marker_spec}{numeric vector, specification scores for corresponding genes in marker_expression}
}
\value{
numeric, index of cell identity score for this expression/spec score combination
}
\description{
Compute ICI score for a given cell type (Internal)
}
