% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{compute_ici_scores}
\alias{compute_ici_scores}
\title{Compute ICI scores for a set of cells}
\usage{
compute_ici_scores(expression_data, spec_table, sig = FALSE,
  n_iterations = 1000, information_level = 20, min_spec_score = 0.15)
}
\arguments{
\item{expression_data}{data frame with at least one column, "Locus", and
additional columns containing expression values for each cell.}

\item{spec_table}{a data.frame containing at least 4 columns:
\itemize{
  \item Locus
  \item Cell_Type
  \item spec (the Specificity score calculated using compute_spec_table)
  \item mean_expr (the mean expression level for that locus/cell type
  combination used to generated spec scores)
}}

\item{sig}{logical, whether to compute significance scores for ici values}

\item{n_iterations}{integer, number of iterations for empirical p-value computation}

\item{information_level}{numeric, cumulative information score when determining marker loci}

\item{min_spec_score}{minimum useful spec score for determining marker loci}
}
\value{
a data frame containing ici values and p-values for all cell/cell-type combinations.
}
\description{
Compute ICI scores for a set of cells
}
\examples{
## Compute ICI scores for a set of test data:

# Train a spec table given initial expression data
spec_table <- compute_spec_table(expression_data = test_spec,
                                 bin_method = "Efroni",
                                 mean_method = "median",
                                 l = 5, u = 3)

# For new cell profiles, compute ICI scores
ici_scores <- compute_ici_scores(expression_data = test_ici,
                                 spec_table = spec_table,
                                 n_iterations = 1000,
                                 sig = TRUE,
                                 information_level = 10,
                                 min_spec_score = 0)
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification
of cell identity from single-cell gene expression profiles". Genome Biology
16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy
biological systems". Nucl. Acids. Res. 39(21)
}
