% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spec_tools.R
\name{compute_mean_expr}
\alias{compute_mean_expr}
\title{Comptue mean expression for an `expression_data` object}
\usage{
compute_mean_expr(expression_data, method = "Efroni", ...)
}
\arguments{
\item{expression_data}{data frame (or tibble) containing the following
columns for a single locus

\itemize{
  \item Expression (the normalized expression value, not log-transformed)
  \item Cell_Type (the cell type from which the expression data came from)
  \item binsize (optional; required if `method = "Efroni"`)
}}

\item{method}{Which method to use when computing the mean. Defaults to
"Efroni", but can be another function. If set to something other than
"Efroni", a function, or "median", defaults to computing the median
expression for each cell type. Must return one or more summarized expression
values for each cell type.}

\item{...}{Parameters to be supplied to the function defined in `method`.}
}
\value{
Summarized expression values for each cell type.
}
\description{
Comptue mean expression for an `expression_data` object
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification of cell identity from single-cell gene expression profiles". Genome Biology 16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy biological systems". Nucl. Acids. Res. 39(21)
}
