% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spec_tools.R
\name{compute_spec}
\alias{compute_spec}
\title{Compute Specification Score}
\usage{
compute_spec(binned_expression_data)
}
\arguments{
\item{binned_expression_data}{data frame (or tibble) containing the following
columns for a single locus

\itemize{
  \item Expression (the normalized expression value, not log-transformed)
  \item Cell_Type (the cell type from which the expression data came from)
  \item bin (which expression bin the expression for this locus falls in per dataset)
}}
}
\value{
A tibble with cell-type specific spec scores for this locus. Supports
negative spec scores (determined by whether the mean expression for that cell
type is above/below the mean expression over all data sets.)
}
\description{
For expression data for a given locus, calculates the
specificity score, as outlined in Birnbaum, et al., 2011. Supports any number
of bins, which MUST be specified in the `bin` column of the input object.
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification of cell identity from single-cell gene expression profiles". Genome Biology 16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy biological systems". Nucl. Acids. Res. 39(21)
}
