% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efroni_specific_functions.R
\name{efroni_bin}
\alias{efroni_bin}
\title{Optimize the selection of bin size for defining two expression classes
in spec score computation.}
\usage{
efroni_bin(expression_data, l = 10, u = 3, use_original = F)
}
\arguments{
\item{expression_data}{data frame (or tibble) containing the following
columns for a single locus

\itemize{
  \item Expression (the normalized expression value, not log-transformed)
  \item Cell_Type (the cell type from which the expression data came from)
}}

\item{l}{integer, the number of cuts to perform when binning expression values (see Efroni, et al., 2016)}

\item{u}{integer, the maximum bin to consider as a background bin (0 < u < l)}

\item{use_original}{logical, whether to use the original method of computing
background bins (ignores minimum values). FALSE by default.}
}
\description{
Optimize the selection of bin size for defining two expression classes
in spec score computation.
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification
of cell identity from single-cell gene expression profiles". Genome Biology
16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy
biological systems". Nucl. Acids. Res. 39(21)
}
