% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efroni_specific_functions.R
\name{efroni_get_background_bin}
\alias{efroni_get_background_bin}
\title{Find the background bin in an expression vector}
\usage{
efroni_get_background_bin(expression, expression_thresh, l,
  use_original = F)
}
\arguments{
\item{expression}{a numeric vector of expression values}

\item{expression_thresh}{the threhold for determining whether a bin is in the background}

\item{l}{integer, number of expression bins}

\item{use_original}{logical, whether to use the original method of computing
background bins (ignores minimum values). FALSE by default.}
}
\value{
an integer corresponding to the "background bin". If this bin is
undetermined (no clear grouping of data), returns the maximum bin number (l)
}
\description{
Implements modified Efroni method for determining which bin
corresponds to background expression. This method is used internally within
the `bin_expression_data`` function.
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification
of cell identity from single-cell gene expression profiles". Genome Biology
16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy
biological systems". Nucl. Acids. Res. 39(21)
}
