% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efroni_specific_functions.R
\name{efroni_get_binsizes}
\alias{efroni_get_binsizes}
\title{Get binsize range for optimization}
\usage{
efroni_get_binsizes(expression, background_bin, l)
}
\arguments{
\item{expression}{numeric, expression vector}

\item{background_bin}{pre-computed background bin. If NA, whole expression
range is considered.}

\item{l}{number of cuts (bins). If < 1, whole expression range is considered.}
}
\value{
a vector of 50 bin sizes to iterate over.
}
\description{
If the number of cuts to test are specified, compute the range
of possible bin sizes as close to the transition to "background" expression
as possible. If the number of cuts is not specified, attempts to optimize the binsize across the whole range of gene expression values (50 evenly spaced binsizes).
}
