% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efroni_specific_functions.R
\name{efroni_mean}
\alias{efroni_mean}
\title{Compute mean expression for each cell type (Efroni method)}
\usage{
efroni_mean(expression_data)
}
\arguments{
\item{expression_data}{data frame (or tibble) minimally containing the
following columns for a single locus

\itemize{
  \item Expression (the normalized expression value, not log-transformed)
  \item Cell_Type (the cell type from which the expression data came from)
  \item binsize (the optimized binsize used to compute expression bins)
}}
}
\value{
a data frame (or tibble) containing the summarized mean and
median expression values for each cell type.
}
\description{
Uses the Efroni method to compute mean expression for each cell
type present in `expression_data`. Here, it computes the mean expression,
divided by the "binsize" used to classify expression.
}
\references{
Ifroni, E., Ip, PL., Nawy, T., Mello, A., Birnbaum, KD. 2015. "Quantification of cell identity from single-cell gene expression profiles". Genome Biology 16(9)

Birnbaum, KD. and Kussell, E. 2011. "Measuring cell identity in noisy biological systems". Nucl. Acids. Res. 39(21)
}
