% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{extract_signal_from_info_data}
\alias{extract_signal_from_info_data}
\title{Extract mean, maximimum ICI signal per cell per information level (internal)}
\usage{
extract_signal_from_info_data(info_data)
}
\arguments{
\item{info_data}{information data structure returned by gather_information_level_data}
}
\value{
a data frame with information level/ICI signal data
}
\description{
Extract mean, maximimum ICI signal per cell per information level (internal)
}
