% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{extract_var_from_info_data}
\alias{extract_var_from_info_data}
\title{Extract variation between information levels for all cells (internal)}
\usage{
extract_var_from_info_data(info_data)
}
\arguments{
\item{info_data}{information data structure returned by gather_information_level_data}
}
\value{
a data frame with information level/ICI variation information
}
\description{
Extract variation between information levels for all cells (internal)
}
