% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{find_ici_markers}
\alias{find_ici_markers}
\title{Find Markers for ICI score computation}
\usage{
find_ici_markers(spec_table, information_level, min_spec_score = 0.15,
  transform_spec = TRUE)
}
\arguments{
\item{spec_table}{a data.frame containing at least 4 columns:
\itemize{
  \item Locus
  \item Cell_Type
  \item spec (the Specificity score calculated using compute_spec_table)
  \item mean_expr (the mean expression level for that locus/cell type
  combination used to generated spec scores)
}}

\item{information_level}{numeric, the maximum cumulative information level
used to compute the ICI score}

\item{min_spec_score}{numeric, minimum useful specification score}

\item{transform_spec}{logical, whether to convert spec scores below
min_spec_score to 0 (default: TRUE)}
}
\value{
Transforms the input spec_table with two additional columns:
\itemize{
  \item cumulative_sum, the cumulative sum of spec scores descending from the
  highest to lowest (with mean_expr used to break ties)
  \item is_marker whether the cumulative sum for that locus is below the
  information threshold specified by information_level
}
}
\description{
Find Markers for ICI score computation
}
