% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ici.R
\name{gather_information_level_data}
\alias{gather_information_level_data}
\title{Find ICI scores over a range of information levels}
\usage{
gather_information_level_data(expression_data, spec_table,
  information_range = seq(0, 100, 0.5), min_spec_score = 0.15,
  n_samples = 100)
}
\arguments{
\item{expression_data}{data frame with at least one column, "Locus", and
additional columns containing expression values for each cell.}

\item{spec_table}{a data.frame containing at least 4 columns:
\itemize{
  \item Locus
  \item Cell_Type
  \item spec (the Specificity score calculated using compute_spec_table)
  \item mean_expr (the mean expression level for that locus/cell type
  combination used to generated spec scores)
}}

\item{information_range}{numeric vector, range of information scores to loop over}

\item{min_spec_score}{minimum useful information score (default, 0.15)}

\item{n_samples}{number of sub-samples to take (useful for limiting computational resources required; default 100)}
}
\value{
a data frame with ici scores for all sampled cells, all cell types,
at all information levels in information_range.
}
\description{
Find ICI scores over a range of information levels
}
