% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spec_tools.R
\name{melted_to_mat}
\alias{melted_to_mat}
\title{Convert melted data.frame (or tibble) to matrix}
\usage{
melted_to_mat(df, var1, var2, indexvar)
}
\arguments{
\item{df}{the data.frame}

\item{var1}{character, Which variable in df has the new column names}

\item{var2}{character, Which variable in df has the new values}

\item{indexvar}{character, which variable in df has the new rownames}
}
\value{
a matrix
}
\description{
internal helper function for testing
}
