% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ici}
\alias{test_ici}
\title{Random toy dataset simulating cell profiles}
\format{A dataset of 15 cells, with the same 14 loci as described in
\code{test_spec}. Conforms to the expected expression_data format, with 3
columns:

\describe{
  \item{Locus}{Locus identifer (chr)}
  \item{Expression}{Expression of this locus in a given dataset (dbl)}
  \item{Cell}{cell type of the data set (chr)}
}

Can be used to test the compute_ici_scores function.}
\usage{
test_ici
}
\description{
Random toy dataset simulating cell profiles
}
\keyword{datasets}
