% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_spec}
\alias{test_spec}
\title{Random toy dataset simulating marker loci}
\format{A dataset of 10 loci, measured across 5 hypothetical cell types (5
replicates per cell type). Conforms to the expected expression_data format,
with 4 columns:

\describe{
  \item{Locus}{Locus identifer (chr)}
  \item{Expression}{Expression of this locus in a given dataset (dbl)}
  \item{Cell_Type}{cell type of the data set (chr)}
  \item{Sample_Name}{name of the "sample" this locus/Cell_Type measurement originated from}
}

Can be used to test the compute_spec_table function.}
\usage{
test_spec
}
\description{
Random toy dataset simulating marker loci
}
\keyword{datasets}
